remove sections
