--- app/sections/Search/index.ts
@@ -1,2 +1,0 @@
-import Search from "./Search";
-export default Search;
